implement next_range and prev_range from the api
